requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
nltk==3.8.1
pyspellchecker==0.8.1
python-dotenv==1.0.1
python-telegram-bot==21.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.1
textblob==0.18.0.post0
tqdm==4.66.2
urllib3==2.2.1
